requirements_api.txt
PyYAML==5.3.1
PyQt5==5.14.2
PyQt5-tools==5.14.2.2
PyQt5-sip==12.7.2
Pillow==6.2.2
black==19.3b0
clipboard==0.0.4
coloredlogs==14.0
fernet==1.0.1
flake8==3.8.3
flup==1.0.3
fpdf2==2.0.5
ninja==1.10.0.post1
odfpy
ply
psycopg2
pyroma==2.6
sqlalchemy==1.4.0b1
sqlparse==0.3.1
twine==3.1.1
keyrings.alt>=3.4.0
watchdog>=0.10.2
python-barcode
pyfiglet
python-daemon
pdf2image
xmlsig
lxml
pyOpenSSL
xades
qrcode[pil]